+++ pylinkage/visualizer.py
@@ -11,7 +11,7 @@ import matplotlib.pyplot as plt
 import matplotlib.animation as anim
 from .utility import movement_bounding_box
 from .interface.exceptions import UnbuildableError
-from .interface import Crank, Fixed, Static, Pivot
+from .interface import Crank, Fixed, Static, Pivot, Revolute
 
 # List of animations
 ANIMATIONS = []
@@ -21,7 +21,8 @@ COLOR_SWITCHER = {
     Static: 'k',
     Crank: 'g',
     Fixed: 'r',
-    Pivot: 'b'
+    Pivot: 'b',
+    Revolute: 'b'
 }