+++ tests/test_joints.py
@@ -10,7 +10,7 @@ import unittest
 import math
 from pylinkage import UnbuildableError
 from pylinkage.interface.joint import Fixed
-from pylinkage.interface.revolute_joint import Pivot
+from pylinkage.interface.revolute_joint import Revolute, Pivot
 
 
 class TestPivot(unittest.TestCase):
@@ -52,6 +52,52 @@ class TestPivot(unittest.TestCase):
         """Test system almost breaking."""
         pivot2 = Pivot(0, 3)
         pivot3 = Pivot(
+            y=1, joint0=self.pivot1, joint1=pivot2,
+            distance0=1, distance1=1
+        )
+        with self.assertRaises(UnbuildableError):
+            pivot3.reload()
+
+
+class TestRevolute(unittest.TestCase):
+    """Test Revolute Joint."""
+
+    pivot1 = Revolute(0, 0)
+
+    def test_buildable(self):
+        """Upper intersect test."""
+        pivot2 = Revolute(1, 0)
+        pivot3 = Revolute(
+            y=1, joint0=self.pivot1, joint1=pivot2,
+            distance0=1, distance1=1
+        )
+        pivot3.reload()
+        self.assertEqual(pivot3.coord(), (.5, math.sqrt(.75)))
+
+    def test_under_intersect(self):
+        """Under intersect test."""
+        pivot2 = Revolute(1, 0)
+        pivot3 = Revolute(
+            y=-1, joint0=self.pivot1, joint1=pivot2,
+            distance0=1, distance1=1
+        )
+        pivot3.reload()
+        self.assertEqual(pivot3.coord(), (.5, -math.sqrt(.75)))
+
+    def test_limit_intersect(self):
+        """Test system almost breaking."""
+        pivot2 = Revolute(2, 0)
+        pivot3 = Revolute(
+            y=1, joint0=self.pivot1, joint1=pivot2,
+            distance0=1, distance1=1
+        )
+        pivot3.reload()
+        self.assertEqual(pivot3.coord(), (1, 0))
+
+    def test_no_intersect(self):
+        """Test system almost breaking."""
+        pivot2 = Revolute(0, 3)
+        pivot3 = Revolute(
             y=1, joint0=self.pivot1, joint1=pivot2,
             distance0=1, distance1=1
         )